Refactor: Standardize user balance data type to float
This commit changes the data type used for user balances from `double` to `float` across multiple files. This change aims to optimize memory usage and potentially improve performance where high precision is not strictly necessary for monetary values.

**Key Changes:**

-   **`model/User.java`:**
    -   The `userBalance` field type changed from `double` to `float`.
    -   Getter `getUserBalance()` now returns `float`.
    -   Setter `setUserBalance()` now accepts `float`.
-   **`util/UserSession.java`:**
    -   The `updateBalance(Context context)` method now returns `float` and calls `MySQLConnector.getUserBalance()` which also returns `float`.
-   **`util/MySQLConnector.java`:**
    -   The `getUserBalance(String email, Context context)` method now returns `float`.
    -   The local `balance` variable within this method is now `float`.
    -   Default return value in case of connection failure changed to `0.0F`.
-   **`util/ResultSetParser.java`:**
    -   When parsing the user data from `ResultSet`, `resultSet.getDouble(2)` was changed to `resultSet.getFloat(2)` to retrieve the user balance as a `float`.
-   **`controller/DashboardController.java`:**
    -   In `ProfileController.handleBalanceRefresh()`:
        -   Variables `previousBalance` and `updatedBalance` are now of type `float`.
        -   The call to `UserSession.updateBalance(context)` now expects and receives a `float`.
-   **`view/fragment/HomeDashboardFragment.java`:**
    -   The `displayBalance(double balance)` method signature changed to `displayBalance(float balance)`.
-   **`view/Iface/IHomeDashboardView.java`:**
    -   The `displayBalance(double balance)` method signature in the interface changed to `displayBalance(float balance)`.

+++ app/src/main/java/com/example/mandelamoney/view/fragment/HomeDashboardFragment.java
@@ -131,7 +131,7 @@ public class HomeDashboardFragment extends Fragment implements IHomeDashboardVie
     }
 
     @Override
-    public void displayBalance(double balance) {
+    public void displayBalance(float balance) {
         String display = "R " + String.format(Locale.getDefault(), "%.2f", balance);
         txtBalance.setText(display);
     }
